src/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pyasn1==0.4.4
python-ldap==3.1.0
Werkzeug==0.14.1
